requirements.txt
langchain
langchain_anthropic
langchain_chroma
langchain_cohere
langchain_community
langchain_huggingface
langchain_openai
langchain_text_splitters
python-dotenv
transformers
lxml
pypdf
pandas
csv2pdf
google-generativeai
azure-ai-formrecognizer
dataclasses
